+++ src/components/manifold-product-details/manifold-product-details.tsx
@@ -1,5 +1,6 @@
 import { h, Component, Prop } from '@stencil/core';
-import { Catalog } from '../../types/catalog';
+
+import { Product } from '../../types/graphql';
 import skeletonProduct from '../../data/product';
 import logger from '../../utils/logger';
 
@@ -9,37 +10,43 @@ import logger from '../../utils/logger';
   shadow: true,
 })
 export class ManifoldProductDetails {
-  @Prop() product?: Catalog.Product;
+  @Prop() product?: Product;
 
   @logger()
   render() {
-    const {
-      body: { tagline, value_props, images = [] },
-    } = this.product || skeletonProduct;
-
     if (this.product) {
+      const { tagline, valueProps, screenshots } = this.product;
       return (
         <div>
           <h1 class="title" itemprop="tagline">
             <span class="tagline">{tagline}</span>
           </h1>
           <ul class="value-prop-list" itemprop="description">
-            {value_props.map(({ body, header }) => (
-              <li class="value-prop" key={header}>
-                <h3>{header}</h3>
-                <p>{body}</p>
-              </li>
-            ))}
+            {valueProps &&
+              valueProps.map(({ body, header }) => (
+                <li class="value-prop" key={header}>
+                  <h3>{header}</h3>
+                  <p>{body}</p>
+                </li>
+              ))}
           </ul>
-          {images.length > 0 && (
-            <manifold-image-gallery images={images}>Screenshots</manifold-image-gallery>
+          {screenshots && screenshots.length > 0 && (
+            <manifold-image-gallery
+              images={[...screenshots]
+                .sort((image1, image2) => image1.order - image2.order)
+                .map(image => image.url)}
+            >
+              Screenshots
+            </manifold-image-gallery>
           )}
         </div>
       );
     }
 
     // 💀
-
+    const {
+      body: { tagline, value_props },
+    } = skeletonProduct;
     return (
       <div>
         <h1 class="title" itemprop="tagline">